Tile prediction is working on training level
The current approach of learning exact tile pattern consequences won't generalise all that well, but predictions of next Sokoban board state is working well now based on interactive tests.  However, Score prediction is NOT working well, because most of the time the tile patterns observe no reward, and when the occasional reward does occur, it's hard to overturn that.